requirements.txt
antlr4-python3-runtime==4.8
appdirs==1.4.4
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
docker-pycreds==0.4.0
filelock==3.9.0
fonttools==4.25.0
gitdb==4.0.10
GitPython==3.1.34
hydra-core==1.0.6
idna==3.4
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
Jinja2==3.1.2
kiwisolver @ file:///C:/b/abs_88mdhvtahm/croot/kiwisolver_1672387921783/work
MarkupSafe==2.1.2
matplotlib @ file:///C:/b/abs_085jhivdha/croot/matplotlib-suite_1693812524572/work
mpmath==1.2.1
munkres==1.1.4
networkx==3.0
numpy @ file:///D:/bld/numpy_1653325737787/work
omegaconf==2.0.6
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pathtools==0.1.2
Pillow==9.3.0
ply==3.11
protobuf==4.24.2
psutil==5.9.5
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PyQt5==5.15.7
PyQt5-sip @ file:///C:/Windows/Temp/abs_d7gmd2jg8i/croots/recipe/pyqt-split_1659273064801/work/pyqt_sip
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
PyWavelets @ file:///C:/b/abs_a8r4b1511a/croot/pywavelets_1670425185881/work
PyYAML==6.0.1
requests==2.28.1
sentry-sdk==1.30.0
setproctitle==1.3.2
sip @ file:///C:/Windows/Temp/abs_b8fxd17m2u/croots/recipe/sip_1659012372737/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sympy==1.11.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado @ file:///C:/b/abs_61jhmrrua1/croot/tornado_1690848767317/work
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.15.10
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
